feat: Add ability to package LLRT as a layer

+++ src/llrt-function.ts
@@ -1,8 +1,9 @@
 import { posix } from 'path';
 import { CfnResource } from 'aws-cdk-lib';
-import { Architecture, Runtime, RuntimeFamily } from 'aws-cdk-lib/aws-lambda';
+import { Architecture, Code, ILayerVersion, LayerVersion, Runtime, RuntimeFamily } from 'aws-cdk-lib/aws-lambda';
 import { ICommandHooks, NodejsFunction, NodejsFunctionProps, OutputFormat } from 'aws-cdk-lib/aws-lambda-nodejs';
 import { Construct } from 'constructs';
+import { execSync } from 'child_process';
 
 /**
  * The type of LLRT binary to use.
@@ -46,6 +47,25 @@ export interface LlrtFunctionProps extends NodejsFunctionProps {
    * @default - If this option is not provided, the LLRT binary is downloaded from GitHub and cached in the .tmp directory.
    */
   readonly llrtBinaryPath?: string;
+
+  /**
+   * If `true` then the LLRT runtime will be built in a layer that can be shared amongst
+   * other `LLrtFunction`s that utilise the same `LlrtBinaryType`.
+   * 
+   * @default - false
+   */
+  readonly llrtLayer?: boolean;
+}
+
+const pathLayers: { [binaryPath: string]: ILayerVersion[]} = {};
+
+const buildLayer = (scope: Construct, binaryPath: string): ILayerVersion[] => {
+  if (!pathLayers[binaryPath]) {
+    pathLayers[binaryPath] = [new LayerVersion(scope, `llrt-layer${Object.keys(pathLayers).length}`, {
+      code: Code.fromAsset(posix.dirname(binaryPath)),
+    })];
+  }
+  return pathLayers[binaryPath];
 }
 
 export class LlrtFunction extends NodejsFunction {
@@ -114,25 +134,30 @@ export class LlrtFunction extends NodejsFunction {
       version == 'latest'
         ? `https://github.com/awslabs/llrt/releases/latest/download/${binaryName}.zip`
         : `https://github.com/awslabs/llrt/releases/download/${version}/${binaryName}.zip`;
-    const cacheDir = `.tmp/llrt/${version}/${arch}/${binaryType}`;
+    const cacheDir = posix.join(__dirname, `/../.tmp/llrt/${version}/${arch}/${binaryType}`);
 
-    const { commandHooks: originalCommandHooks, ...otherBundlingProps } = props.bundling ?? {};
-    const afterBundlingCommandHook: ICommandHooks['afterBundling'] = (i, o) => !props.llrtBinaryPath ? [
-      // Download llrt binary from GitHub release and cache it
-      `if [ ! -e ${posix.join(i, cacheDir, 'bootstrap')} ]; then
-        mkdir -p ${posix.join(i, cacheDir)}
-        cd ${posix.join(i, cacheDir)}
+    if (!props.llrtBinaryPath) {
+      execSync(`if [ ! -e ${posix.join(cacheDir, 'bootstrap')} ]; then
+        mkdir -p ${cacheDir}
+        cd ${cacheDir}
         curl -L -o llrt_temp.zip ${binaryUrl}
         unzip llrt_temp.zip
         rm -rf llrt_temp.zip
         cd -
-       fi`,
-      `cp ${posix.join(i, cacheDir, 'bootstrap')} ${o}`,
-    ] : [`cp ${posix.join(i, props.llrtBinaryPath)} ${posix.join(o, 'bootstrap')}`];
+      fi`);
+    }
+
+    const binaryPath = !props.llrtBinaryPath ? posix.join(cacheDir, 'bootstrap') : props.llrtBinaryPath;
+
+    const { commandHooks: originalCommandHooks, ...otherBundlingProps } = props.bundling ?? {};
+    const afterBundlingCommandHook: ICommandHooks['afterBundling'] = (_i, o) => props.llrtLayer ? [] : [`cp ${binaryPath} ${posix.join(o, 'bootstrap')}`];
+
+    const layers = props.llrtLayer ? buildLayer(scope, binaryPath) : [];
 
     super(scope, id, {
       // set this to remove an unnecessary environment variable.
       awsSdkConnectionReuse: false,
+      layers,
       // set this to remove a warning about runtime. we use al2023 runtime anyway.
       runtime: new Runtime('nodejs20.x', RuntimeFamily.NODEJS),
       ...props,
